backend/requirements.txt
fastapi
uvicorn
pyjwt
kubernetes
kopf
requests
pydantic
